3.2 Build 1211
git-svn-id: https://plugins.svn.wordpress.org/easyrecipe/trunk@655200 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -1,5 +1,11 @@
 <?php
-delete_option("EasyRecipe");
-$data = http_build_query(array ('action' => 'uninstall', 'site' => get_site_url()));
-$fp = @fopen("http://www.easyrecipeplugin.com/installed.php?$data", "r");
+if (!function_exists('add_action')) {
+    echo "Hi there!  I'm just a plugin, not much I can do when called directly.";
+    exit();
+}
 
+if (current_user_can('edit_plugins')) {
+    delete_option("EasyRecipe");
+    $data = http_build_query(array('action' => 'uninstall', 'site' => get_site_url()));
+    $fp = @fopen("http://www.easyrecipeplugin.com/installed.php?$data", "r");
+}
